// web/src/lib/api/generated/models/datastoreColumnResource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * KilasFlow API
 * Embeddable workflow automation engine. Every operation available in the editor is available here: the canvas is a client of this API, not the owner of workflow state.
 * OpenAPI spec version: 0.1.0-dev
 */

export interface DatastoreColumnResource {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  name: string;
  type: string;
}